fix: remove mimalloc from source
Too bad, under Flatpak or even under Windows, this library can cause the program to crash. The reason for this is that there are some places that do not override the memory allocation method, but this does not guarantee that the mimalloc library will be loaded first in some environments (especially Flatpak) so I think it would be better to let the user install it themselves rather than include this library. Then override all mallocs as needed, which is easy to do. Sacrificing some of the security of the default state for the stability of the program feels necessary to me.