chore: improves packaging and RPATH handling
Refactors RPATH handling for better packaging support,
especially for DEB and RPM packages.

This change introduces a dedicated option to strip RPATH,
ensuring binaries are correctly packaged. When building
an app for packaging, it's now standard to strip RPATH.
The logic for setting CMAKE_INSTALL_RPATH is also adjusted
to account for the strip RPATH option and Xcode builds.

Additionally, fixes an issue where the packaging install
prefix was not correctly set in src/CMakeLists.txt.